Version 6.7.159.1 (cherry-pick)
Merged 2746f658b70a890675e2ba919a876471d4ee0047

Revert "Force proper code aligment in wasm serializer."

R=hablich@chromium.org

Cq-Include-Trybots: luci.chromium.try:linux_chromium_rel_ng
Reviewed-on: https://chromium-review.googlesource.com/979448
Cr-Commit-Position: refs/heads/6.7.159@{#2}
Cr-Branched-From: 62515936747e2cf1365260e820ed87f4bb4c47dc-refs/heads/master@{#52180}

+++ src/wasm/wasm-code-manager.h
@@ -100,7 +100,7 @@ class V8_EXPORT_PRIVATE WasmCode final {
   // kOther is used if we have WasmCode that is neither
   // liftoff- nor turbofan-compiled, i.e. if Kind is
   // not a kFunction.
-  enum Tier : int32_t { kLiftoff, kTurbofan, kOther };
+  enum Tier : int8_t { kLiftoff, kTurbofan, kOther };
 
   Vector<byte> instructions() const { return instructions_; }
   Vector<const byte> reloc_info() const {
